Move MS-IME user dictionary importer to gui/base.
This is a series of CLs to drop dependency on iconv (#252).

MS-IME user dictionary importer is only used from src/gui/.
Moving it to src/gui/base/ makes build dependency simpler and possibly
improves build time depending on build configurations and build
targets.

BUG=#252
TEST=unittest
REF_BUG=19010851
REF_CL=87220744
